Refactor splash screen: separate animations into dedicated widgets and enhance structure * REFACTOR SPLASH SCREEN WIDGETS
