+++ lib/features/splash_screen/ui/views/splash_view.dart
@@ -1,6 +1,4 @@
-import 'package:az_task/core/extensions/navigation_extenstion.dart';
-import 'package:az_task/core/images/assets_images.dart';
-import 'package:az_task/core/routes/app_routes.dart';
+import 'package:az_task/features/splash_screen/ui/widgets/splash_view_body.dart';
 import 'package:flutter/material.dart';
 
 class SplashScreen extends StatelessWidget {
@@ -9,102 +7,5 @@ class SplashScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(body: SplashViewBody());
-  }
-}
-
-class SplashViewBody extends StatefulWidget {
-  const SplashViewBody({super.key});
-
-  @override
-  State<SplashViewBody> createState() => _SplashViewBodyState();
-}
-
-class _SplashViewBodyState extends State<SplashViewBody>
-    with SingleTickerProviderStateMixin {
-  late AnimationController animationController;
-  late Animation<Offset> slidingAnimation;
-  late Animation<double> fadeAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    initAnimation();
-    goToHome();
-  }
-
-  void goToHome() {
-    Future.delayed(const Duration(seconds: 60), () {
-      // ignore: use_build_context_synchronously
-      context.pushReplacmentNamed(AppRoutes.onboardingScreen);
-    });
-  }
-
-  @override
-  void dispose() {
-    // Dispose of the animation controller
-    animationController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        FadeTransition(
-          opacity: fadeAnimation,
-
-          child: const Image(image: AssetImage(AssetsImages.logo), height: 150),
-        ),
-        const SizedBox(height: 8),
-        SlidingAnimationText(slidingAnimation: slidingAnimation),
-      ],
-    );
-  }
-
-  void initAnimation() {
-    animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 1),
-    );
-    slidingAnimation = Tween<Offset>(
-      begin: const Offset(0, 2),
-      end: Offset.zero,
-    ).animate(animationController);
-
-    fadeAnimation = Tween<double>(
-      begin: 0.0,
-      end: 1.0,
-    ).animate(animationController);
-
-    // Start the animation
-    animationController.forward();
-  }
-}
-
-class SlidingAnimationText extends StatelessWidget {
-  const SlidingAnimationText({super.key, required this.slidingAnimation});
-
-  final Animation<Offset> slidingAnimation;
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: slidingAnimation,
-      builder: (context, _) {
-        return SlideTransition(
-          position: slidingAnimation,
-          child:  Text(
-            'Welcome to Az Task',
-            textAlign: TextAlign.center,
-            style: TextTheme.of(context).bodyLarge?.copyWith(
-              fontSize: 24,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        );
-      },
-    );
   }
 }
