+++ lib/features/splash_screen/ui/widgets/slide_animation_text.dart
@@ -1,0 +1,28 @@
+
+import 'package:flutter/material.dart';
+
+class SlidingAnimationText extends StatelessWidget {
+  const SlidingAnimationText({super.key, required this.slidingAnimation});
+
+  final Animation<Offset> slidingAnimation;
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedBuilder(
+      animation: slidingAnimation,
+      builder: (context, _) {
+        return SlideTransition(
+          position: slidingAnimation,
+          child:  Text(
+            'Welcome to Az Task',
+            textAlign: TextAlign.center,
+            style: TextTheme.of(context).bodyLarge?.copyWith(
+              fontSize: 24,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
